tests/wpt/web-platform-tests/css/css-text-decor/parsing/text-decoration-color-valid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Text Decoration Test: Parsing text-decoration-color with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-3/#text-decoration-color-property">
<meta name="assert" content="text-decoration-color supports the full grammar '<color>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
test_valid_value("text-decoration-color", "currentcolor");
test_valid_value("text-decoration-color", "red");
test_valid_value("text-decoration-color", "rgba(10, 20, 30, 0.4)");
</script>
